python/v1/src/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.6
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==3.4
numpy==1.25.2
opencv-python==4.8.0.74
Pillow==10.0.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.6
